Move test utilities used by themes into the themes directory.
