+++ package.json
@@ -46,6 +46,7 @@
     "gatsby-theme-amaranth": "workspace:*",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-query": "^3.19.1",
     "styled-components": "^5.3.0"
   },
   "devDependencies": {
