+++ themes/advanced/src/components/SEO/tests/index.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import SEO from "../index";
-import { post, config } from "../../../../test/fixtures";
+import { post, config } from "../../../../../test/fixtures";
 
 jest.mock("react", () => ({
   ...jest.requireActual<typeof React>("react"),
