+++ themes/test/fixtures/post.ts
@@ -1,4 +1,4 @@
-import { Post } from "../../src/types";
+import { Post } from "../../advanced/src/types";
 
 const post: Post = {
   body: 'function _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }\n\nfunction _objectWithoutProperties(source, excluded) { if (source == null) return {}; var target = _objectWithoutPropertiesLoose(source, excluded); var key, i; if (Object.getOwnPropertySymbols) { var sourceSymbolKeys = Object.getOwnPropertySymbols(source); for (i = 0; i < sourceSymbolKeys.length; i++) { key = sourceSymbolKeys[i]; if (excluded.indexOf(key) >= 0) continue; if (!Object.prototype.propertyIsEnumerable.call(source, key)) continue; target[key] = source[key]; } } return target; }\n\nfunction _objectWithoutPropertiesLoose(source, excluded) { if (source == null) return {}; var target = {}; var sourceKeys = Object.keys(source); var key, i; for (i = 0; i < sourceKeys.length; i++) { key = sourceKeys[i]; if (excluded.indexOf(key) >= 0) continue; target[key] = source[key]; } return target; }\n\n/* @jsxRuntime classic */\n\n/* @jsx mdx */\nvar _frontmatter = {\n  "title": "Big Test",\n  "cover": "5.jpg",\n  "description": "Decsription",\n  "category": "moar",\n  "datePublished": "2018-03-01",\n  "dateModified": "2018-03-01",\n  "slug": "Big-Sample_Test",\n  "tags": ["test", "huge"]\n};\nvar layoutProps = {\n  _frontmatter: _frontmatter\n};\nvar MDXLayout = "wrapper";\nreturn function MDXContent(_ref) {\n  var components = _ref.components,\n      props = _objectWithoutProperties(_ref, ["components"]);\n\n  return mdx(MDXLayout, _extends({}, layoutProps, props, {\n    components: components,\n    mdxType: "MDXLayout"\n  }), mdx("h1", {\n    "id": "mdx-test",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h1",\n    "href": "#mdx-test",\n    "aria-label": "mdx test permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "MDX Test"), mdx("p", null, "Here\'s a color picker!"), , mdx("h2", {\n    "id": "note-this-post-is-based-on-markdown-cheatsheet-and-is-meant-to-test-styling-of-markdown-generated-documents",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#note-this-post-is-based-on-markdown-cheatsheet-and-is-meant-to-test-styling-of-markdown-generated-documents",\n    "aria-label": "note this post is based on markdown cheatsheet and is meant to test styling of markdown generated documents permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "NOTE: This \\"post\\" is based on ", mdx("a", {\n    parentName: "h2",\n    "href": "https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "Markdown Cheatsheet"), " and is meant to test styling of Markdown generated documents."), mdx("p", null, "This is intended as a quick reference and showcase. For more complete info, see ", mdx("a", {\n    parentName: "p",\n    "href": "http://daringfireball.net/projects/markdown/",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "John Gruber\'s original spec"), " and the ", mdx("a", {\n    parentName: "p",\n    "href": "http://github.github.com/github-flavored-markdown/",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "Github-flavored Markdown info page"), "."), mdx("p", null, "This cheatsheet is specifically ", mdx("em", {\n    parentName: "p"\n  }, "Markdown Here\'s"), " version of Github-flavored Markdown. This differs slightly in styling and syntax from what Github uses, so what you see below might vary a little from what you get in a ", mdx("em", {\n    parentName: "p"\n  }, "Markdown Here"), " email, but it should be pretty close."), mdx("p", null, "You can play around with Markdown on our ", mdx("a", {\n    parentName: "p",\n    "href": "http://www.markdown-here.com/livedemo.html",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "live demo page"), "."), mdx("p", null, "(If you\'re not a Markdown Here user, check out the ", mdx("a", {\n    parentName: "p",\n    "href": "./Markdown-Cheatsheet"\n  }, "Markdown Cheatsheet"), " that is not specific to MDH. But, really, you should also use Markdown Here, because it\'s awesome. ", mdx("a", {\n    parentName: "p",\n    "href": "http://markdown-here.com",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "http://markdown-here.com"), ")"), mdx("h5", {\n    "id": "table-of-contents",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h5",\n    "href": "#table-of-contents",\n    "aria-label": "table of contents permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Table of Contents"), mdx("p", null, mdx("a", {\n    parentName: "p",\n    "href": "#headers"\n  }, "Headers"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#emphasis"\n  }, "Emphasis"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#lists"\n  }, "Lists"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#links"\n  }, "Links"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#images"\n  }, "Images"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#code"\n  }, "Code and Syntax Highlighting"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#tables"\n  }, "Tables"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#blockquotes"\n  }, "Blockquotes"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#html"\n  }, "Inline HTML"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#hr"\n  }, "Horizontal Rule"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#lines"\n  }, "Line Breaks"), "\\n", mdx("a", {\n    parentName: "p",\n    "href": "#videos"\n  }, "YouTube Videos")), mdx("a", {\n    name: "headers"\n  }), mdx("h2", {\n    "id": "headers",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#headers",\n    "aria-label": "headers permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Headers"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "# H1\\n## H2\\n### H3\\n#### H4\\n##### H5\\n###### H6\\n\\nAlternatively, for H1 and H2, an underline-ish style:\\n\\nAlt-H1\\n======\\n\\nAlt-H2\\n------"))), mdx("h1", {\n    "id": "h1",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h1",\n    "href": "#h1",\n    "aria-label": "h1 permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "H1"), mdx("h2", {\n    "id": "h2",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#h2",\n    "aria-label": "h2 permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "H2"), mdx("h3", {\n    "id": "h3",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h3",\n    "href": "#h3",\n    "aria-label": "h3 permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "H3"), mdx("h4", {\n    "id": "h4",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h4",\n    "href": "#h4",\n    "aria-label": "h4 permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "H4"), mdx("h5", {\n    "id": "h5",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h5",\n    "href": "#h5",\n    "aria-label": "h5 permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "H5"), mdx("h6", {\n    "id": "h6",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h6",\n    "href": "#h6",\n    "aria-label": "h6 permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "H6"), mdx("p", null, "Alternatively, for H1 and H2, an underline-ish style:"), mdx("h1", {\n    "id": "alt-h1",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h1",\n    "href": "#alt-h1",\n    "aria-label": "alt h1 permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Alt-H1"), mdx("h2", {\n    "id": "alt-h2",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#alt-h2",\n    "aria-label": "alt h2 permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Alt-H2"), mdx("a", {\n    name: "emphasis"\n  }), mdx("h2", {\n    "id": "emphasis",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#emphasis",\n    "aria-label": "emphasis permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Emphasis"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "Emphasis, aka italics, with *asterisks* or _underscores_.\\n\\nStrong emphasis, aka bold, with **asterisks** or __underscores__.\\n\\nCombined emphasis with **asterisks and _underscores_**.\\n\\nStrikethrough uses two tildes. ~~Scratch this.~~"))), mdx("p", null, "Emphasis, aka italics, with ", mdx("em", {\n    parentName: "p"\n  }, "asterisks"), " or ", mdx("em", {\n    parentName: "p"\n  }, "underscores"), "."), mdx("p", null, "Strong emphasis, aka bold, with ", mdx("strong", {\n    parentName: "p"\n  }, "asterisks"), " or ", mdx("strong", {\n    parentName: "p"\n  }, "underscores"), "."), mdx("p", null, "Combined emphasis with ", mdx("strong", {\n    parentName: "p"\n  }, "asterisks and ", mdx("em", {\n    parentName: "strong"\n  }, "underscores")), "."), mdx("p", null, "Strikethrough uses two tildes. ", mdx("del", {\n    parentName: "p"\n  }, "Scratch this.")), mdx("a", {\n    name: "lists"\n  }), mdx("h2", {\n    "id": "lists",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#lists",\n    "aria-label": "lists permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Lists"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "1. First ordered list item\\n2. Another item\\n  * Unordered sub-list.\\n1. Actual numbers don\'t matter, just that it\'s a number\\n  1. Ordered sub-list\\n4. And another item.\\n\\n   Some text that should be aligned with the above item.\\n\\n* Unordered list can use asterisks\\n- Or minuses\\n+ Or pluses"))), mdx("ol", null, mdx("li", {\n    parentName: "ol"\n  }, mdx("p", {\n    parentName: "li"\n  }, "First ordered list item")), mdx("li", {\n    parentName: "ol"\n  }, mdx("p", {\n    parentName: "li"\n  }, "Another item"), mdx("ul", {\n    parentName: "li"\n  }, mdx("li", {\n    parentName: "ul"\n  }, "Unordered sub-list."))), mdx("li", {\n    parentName: "ol"\n  }, mdx("p", {\n    parentName: "li"\n  }, "Actual numbers don\'t matter, just that it\'s a number"), mdx("ol", {\n    parentName: "li"\n  }, mdx("li", {\n    parentName: "ol"\n  }, "Ordered sub-list"))), mdx("li", {\n    parentName: "ol"\n  }, mdx("p", {\n    parentName: "li"\n  }, "And another item."), mdx("p", {\n    parentName: "li"\n  }, "Some text that should be aligned with the above item."))), mdx("ul", null, mdx("li", {\n    parentName: "ul"\n  }, "Unordered list can use asterisks")), mdx("ul", null, mdx("li", {\n    parentName: "ul"\n  }, "Or minuses")), mdx("ul", null, mdx("li", {\n    parentName: "ul"\n  }, "Or pluses")), mdx("a", {\n    name: "links"\n  }), mdx("h2", {\n    "id": "links",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#links",\n    "aria-label": "links permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Links"), mdx("p", null, "There are two ways to create links."), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "[I\'m an inline-style link](https://www.google.com)\\n\\n[I\'m a reference-style link][Arbitrary case-insensitive reference text]\\n\\n[You can use numbers for reference-style link definitions][1]\\n\\nOr leave it empty and use the [link text itself]\\n\\nURLs and URLs in angle brackets will automatically get turned into links.\\nhttp://www.example.com or <http://www.example.com> and sometimes\\nexample.com (but not on Github, for example).\\n\\nSome text to show that the reference links can follow later.\\n\\n[arbitrary case-insensitive reference text]: https://www.mozilla.org\\n[1]: http://slashdot.org\\n[link text itself]: http://www.reddit.com"))), mdx("p", null, mdx("a", {\n    parentName: "p",\n    "href": "https://www.google.com",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "I\'m an inline-style link")), mdx("p", null, mdx("a", {\n    parentName: "p",\n    "href": "https://www.mozilla.org",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "I\'m a reference-style link")), mdx("p", null, mdx("a", {\n    parentName: "p",\n    "href": "http://slashdot.org",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "You can use numbers for reference-style link definitions")), mdx("p", null, "Or leave it empty and use the ", mdx("a", {\n    parentName: "p",\n    "href": "http://www.reddit.com",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "link text itself")), mdx("p", null, "URLs and URLs in angle brackets will automatically get turned into links.\\n", mdx("a", {\n    parentName: "p",\n    "href": "http://www.example.com",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "http://www.example.com"), " or ", mdx("a", {\n    parentName: "p",\n    "href": "http://www.example.com",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "http://www.example.com"), " and sometimes\\nexample.com (but not on Github, for example)."), mdx("p", null, "Some text to show that the reference links can follow later."), mdx("a", {\n    name: "images"\n  }), mdx("h2", {\n    "id": "images",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#images",\n    "aria-label": "images permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Images"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "Here\'s our logo (hover to see the title text):\\n\\nInline-style:\\n![alt text](https://github.com/adam-p/markdown-here/raw/master/src/common/images/icon48.png \\"Logo Title Text 1\\")\\n\\nReference-style:\\n![alt text][logo]\\n\\n[logo]: https://github.com/adam-p/markdown-here/raw/master/src/common/images/icon48.png \\"Logo Title Text 2\\""))), mdx("p", null, "Here\'s our logo (hover to see the title text):"), mdx("p", null, "Inline-style:\\n", mdx("img", {\n    parentName: "p",\n    "src": "https://github.com/adam-p/markdown-here/raw/master/src/common/images/icon48.png",\n    "alt": "alt text",\n    "title": "Logo Title Text 1"\n  })), mdx("p", null, "Reference-style:\\n", mdx("img", {\n    parentName: "p",\n    "src": "https://github.com/adam-p/markdown-here/raw/master/src/common/images/icon48.png",\n    "alt": "alt text",\n    "title": "Logo Title Text 2"\n  })), mdx("a", {\n    name: "code"\n  }), mdx("h2", {\n    "id": "code-and-syntax-highlighting",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#code-and-syntax-highlighting",\n    "aria-label": "code and syntax highlighting permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Code and Syntax Highlighting"), mdx("p", null, "Code blocks are part of the Markdown spec, but syntax highlighting isn\'t. However, many renderers -- like Github\'s and ", mdx("em", {\n    parentName: "p"\n  }, "Markdown Here"), " -- support syntax highlighting. ", mdx("em", {\n    parentName: "p"\n  }, "Markdown Here"), " supports highlighting for dozens of languages (and not-really-languages, like diffs and HTTP headers); to see the complete list, and how to write the language names, see the ", mdx("a", {\n    parentName: "p",\n    "href": "http://softwaremaniacs.org/media/soft/highlight/test.html",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "highlight.js demo page"), "."), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "Inline `code` has `back-ticks around` it."))), mdx("p", null, "Inline ", mdx("code", {\n    parentName: "p",\n    "className": "language-text"\n  }, "code"), " has ", mdx("code", {\n    parentName: "p",\n    "className": "language-text"\n  }, "back-ticks around"), " it."), mdx("p", null, "Blocks of code are either fenced by lines with three back-ticks ", mdx("code", null, "```"), ", or are indented with four spaces. I recommend only using the fenced code blocks -- they\'re easier and only they support syntax highlighting."), mdx("pre", {\n    lang: "no-highlight"\n  }, mdx("code", null, "```javascript var s = \\"JavaScript syntax highlighting\\"; alert(s); ``` ```python s = \\"Python syntax highlighting\\" print s ``` ``` No language indicated, so no syntax highlighting. But let\'s throw in a <b>tag</b>. ```")), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "javascript"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-javascript"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-javascript"\n  }, mdx("span", {\n    parentName: "code",\n    "className": "token keyword"\n  }, "var"), " s ", mdx("span", {\n    parentName: "code",\n    "className": "token operator"\n  }, "="), " ", mdx("span", {\n    parentName: "code",\n    "className": "token string"\n  }, "\\"JavaScript syntax highlighting\\""), mdx("span", {\n    parentName: "code",\n    "className": "token punctuation"\n  }, ";"), "\\n", mdx("span", {\n    parentName: "code",\n    "className": "token function"\n  }, "alert"), mdx("span", {\n    parentName: "code",\n    "className": "token punctuation"\n  }, "("), "s", mdx("span", {\n    parentName: "code",\n    "className": "token punctuation"\n  }, ")"), mdx("span", {\n    parentName: "code",\n    "className": "token punctuation"\n  }, ";")))), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "python"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-python"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-python"\n  }, "s ", mdx("span", {\n    parentName: "code",\n    "className": "token operator"\n  }, "="), " ", mdx("span", {\n    parentName: "code",\n    "className": "token string"\n  }, "\\"Python syntax highlighting\\""), "\\n", mdx("span", {\n    parentName: "code",\n    "className": "token keyword"\n  }, "print"), " s"))), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "text"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-text"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-text"\n  }, "No language indicated, so no syntax highlighting in Markdown Here (varies on Github).\\nBut let\'s throw in a <b>tag</b>."))), mdx("p", null, "Again, to see what languages are available for highlighting, and how to write those language names, see the ", mdx("a", {\n    parentName: "p",\n    "href": "http://softwaremaniacs.org/media/soft/highlight/test.html",\n    "target": "_blank",\n    "rel": "nofollow noopener noreferrer"\n  }, "highlight.js demo page"), "."), mdx("a", {\n    name: "tables"\n  }), mdx("h2", {\n    "id": "tables",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#tables",\n    "aria-label": "tables permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Tables"), mdx("p", null, "Tables aren\'t part of the core Markdown spec, but they are part of GFM and ", mdx("em", {\n    parentName: "p"\n  }, "Markdown Here"), " supports them. They are an easy way of adding tables to your email -- a task that would otherwise require copy-pasting from another application."), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "Colons can be used to align columns.\\n\\n| Tables        | Are           | Cool  |\\n| ------------- | :-----------: | ----: |\\n| col 3 is      | right-aligned | $1600 |\\n| col 2 is      | centered      | $12   |\\n| zebra stripes | are neat      | $1    |\\n\\nThe outer pipes (|) are optional, and you don\'t need to make the raw Markdown line up prettily. You can also use inline Markdown.\\n\\n| | Markdown | Less           | Pretty     |     |\\n| | ------------- | --------------- | ---------- |------- |\\n| | *Still*   | `renders` | **nicely** | |\\n| | 1               | 2                 | 3          |          |"))), mdx("p", null, "Colons can be used to align columns."), mdx("table", null, mdx("thead", {\n    parentName: "table"\n  }, mdx("tr", {\n    parentName: "thead"\n  }, mdx("th", {\n    parentName: "tr",\n    "align": null\n  }, "Tables"), mdx("th", {\n    parentName: "tr",\n    "align": "center"\n  }, "Are"), mdx("th", {\n    parentName: "tr",\n    "align": "right"\n  }, "Cool"))), mdx("tbody", {\n    parentName: "table"\n  }, mdx("tr", {\n    parentName: "tbody"\n  }, mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, "col 3 is"), mdx("td", {\n    parentName: "tr",\n    "align": "center"\n  }, "right-aligned"), mdx("td", {\n    parentName: "tr",\n    "align": "right"\n  }, "$1600")), mdx("tr", {\n    parentName: "tbody"\n  }, mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, "col 2 is"), mdx("td", {\n    parentName: "tr",\n    "align": "center"\n  }, "centered"), mdx("td", {\n    parentName: "tr",\n    "align": "right"\n  }, "$12")), mdx("tr", {\n    parentName: "tbody"\n  }, mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, "zebra stripes"), mdx("td", {\n    parentName: "tr",\n    "align": "center"\n  }, "are neat"), mdx("td", {\n    parentName: "tr",\n    "align": "right"\n  }, "$1")))), mdx("p", null, "The outer pipes (|) are optional, and you don\'t need to make the raw Markdown line up prettily. You can also use inline Markdown."), mdx("table", null, mdx("thead", {\n    parentName: "table"\n  }, mdx("tr", {\n    parentName: "thead"\n  }, mdx("th", {\n    parentName: "tr",\n    "align": null\n  }, "Markdown"), mdx("th", {\n    parentName: "tr",\n    "align": null\n  }, "Less"), mdx("th", {\n    parentName: "tr",\n    "align": null\n  }, "Pretty"))), mdx("tbody", {\n    parentName: "table"\n  }, mdx("tr", {\n    parentName: "tbody"\n  }, mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, mdx("em", {\n    parentName: "td"\n  }, "Still")), mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, mdx("code", {\n    parentName: "td",\n    "className": "language-text"\n  }, "renders")), mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, mdx("strong", {\n    parentName: "td"\n  }, "nicely"))), mdx("tr", {\n    parentName: "tbody"\n  }, mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, "1"), mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, "2"), mdx("td", {\n    parentName: "tr",\n    "align": null\n  }, "3")))), mdx("a", {\n    name: "blockquotes"\n  }), mdx("h2", {\n    "id": "blockquotes",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#blockquotes",\n    "aria-label": "blockquotes permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Blockquotes"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "> Blockquotes are very handy in email to emulate reply text.\\n> This line is part of the same quote.\\n\\nQuote break.\\n\\n> This is a very long line that will still be quoted properly when it wraps. Oh boy let\'s keep writing to make sure this is long enough to actually wrap for everyone. Oh, you can *put* **Markdown** into a blockquote."))), mdx("blockquote", null, mdx("p", {\n    parentName: "blockquote"\n  }, "Blockquotes are very handy in email to emulate reply text.\\nThis line is part of the same quote.")), mdx("p", null, "Quote break."), mdx("blockquote", null, mdx("p", {\n    parentName: "blockquote"\n  }, "This is a very long line that will still be quoted properly when it wraps. Oh boy let\'s keep writing to make sure this is long enough to actually wrap for everyone. Oh, you can ", mdx("em", {\n    parentName: "p"\n  }, "put"), " ", mdx("strong", {\n    parentName: "p"\n  }, "Markdown"), " into a blockquote.")), mdx("a", {\n    name: "html"\n  }), mdx("h2", {\n    "id": "inline-html",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#inline-html",\n    "aria-label": "inline html permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Inline HTML"), mdx("p", null, "You can also use raw HTML in your Markdown, and it\'ll mostly work pretty well."), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "<dl>\\n  <dt>Definition list</dt>\\n  <dd>Is something people use sometimes.</dd>\\n\\n  <dt>Markdown in HTML</dt>\\n  <dd>Does *not* work **very** well. Use HTML <em>tags</em>.</dd>\\n</dl>"))), mdx("dl", null, mdx("dt", null, "Definition list"), mdx("dd", null, "Is something people use sometimes."), mdx("dt", null, "Markdown in HTML"), mdx("dd", null, "Does *not* work **very** well. Use HTML ", mdx("em", null, "tags"), ".")), mdx("a", {\n    name: "hr"\n  }), mdx("h2", {\n    "id": "horizontal-rule",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#horizontal-rule",\n    "aria-label": "horizontal rule permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Horizontal Rule"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "text"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-text"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-text"\n  }, "Three or more...\\n\\n---\\n\\nHyphens\\n\\n***\\n\\nAsterisks\\n\\n___\\n\\nUnderscores"))), mdx("p", null, "Three or more..."), mdx("hr", null), mdx("p", null, "Hyphens"), mdx("hr", null), mdx("p", null, "Asterisks"), mdx("hr", null), mdx("p", null, "Underscores"), mdx("a", {\n    name: "lines"\n  }), mdx("h2", {\n    "id": "line-breaks",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#line-breaks",\n    "aria-label": "line breaks permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "Line Breaks"), mdx("p", null, "My basic recommendation for learning how line breaks work is to experiment and discover -- hit ", "<", "Enter", ">", " once (i.e., insert one newline), then hit it twice (i.e., insert two newlines), see what happens. You\'ll soon learn to get what you want. \\"Markdown Toggle\\" is your friend."), mdx("p", null, "Here are some things to try out:"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "text"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-text"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-text"\n  }, "Here\'s a line for us to start with.\\n\\nThis line is separated from the one above by two newlines, so it will be a *separate paragraph*.\\n\\nThis line is also a separate paragraph, but...\\nThis line is only separated by a single newline, so it\'s a separate line in the *same paragraph*."))), mdx("p", null, "Here\'s a line for us to start with."), mdx("p", null, "This line is separated from the one above by two newlines, so it will be a ", mdx("em", {\n    parentName: "p"\n  }, "separate paragraph"), "."), mdx("p", null, "This line is also begins a separate paragraph, but...\\nThis line is only separated by a single newline, so it\'s a separate line in the ", mdx("em", {\n    parentName: "p"\n  }, "same paragraph"), "."), mdx("p", null, "(Technical note: ", mdx("em", {\n    parentName: "p"\n  }, "Markdown Here"), " uses GFM line breaks, so there\'s no need to use MD\'s two-space line breaks.)"), mdx("a", {\n    name: "videos"\n  }), mdx("h2", {\n    "id": "youtube-videos",\n    "style": {\n      "position": "relative"\n    }\n  }, mdx("a", {\n    parentName: "h2",\n    "href": "#youtube-videos",\n    "aria-label": "youtube videos permalink",\n    "className": "anchor before"\n  }, mdx("svg", {\n    parentName: "a",\n    "aria-hidden": "true",\n    "focusable": "false",\n    "height": "16",\n    "version": "1.1",\n    "viewBox": "0 0 16 16",\n    "width": "16"\n  }, mdx("path", {\n    parentName: "svg",\n    "fillRule": "evenodd",\n    "d": "M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"\n  }))), "YouTube Videos"), mdx("p", null, "They can\'t be added directly but you can add an image with a link to the video like this:"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "<a href=\\"http://www.youtube.com/watch?feature=player_embedded&v=8AkLfYOgIrE\\n\\" target=\\"_blank\\"><img src=\\"http://img.youtube.com/vi/8AkLfYOgIrE/0.jpg\\"\\nalt=\\"IMAGE ALT TEXT HERE\\" width=\\"240\\" height=\\"180\\" border=\\"10\\" /></a>"))), mdx("p", null, "Or, in pure Markdown, but losing the image sizing and border:"), mdx("div", {\n    "className": "gatsby-highlight",\n    "data-language": "no-highlight"\n  }, mdx("pre", {\n    parentName: "div",\n    "className": "language-no-highlight"\n  }, mdx("code", {\n    parentName: "pre",\n    "className": "language-no-highlight"\n  }, "[![IMAGE ALT TEXT HERE](http://img.youtube.com/vi/YOUTUBE_VIDEO_ID_HERE/0.jpg)](http://www.youtube.com/watch?v=YOUTUBE_VIDEO_ID_HERE)"))));\n}\n;\nMDXContent.isMDXComponent = true;',
